Add configuration options to disable plugin file remapping
In bukkit.yml, you can now control plugin remapping in mcpc.plugin-settings,
either globally or per-plugin (keyed by the plugin's filename, before '.'):

remap-plugin-file: if true, the plugin is run through SrgTools ApplySrg,
creating a plugin file of the same name but prefixed with 'ported_'. If
the plugin does not function as expected, you can disable this option and
try with the original version of the plugin.