New plugin remapper using SpecialSource. Closes #12.
Features:
* In-memory remapping: the plugin is remapped during load, instead of
rewritten to disk. Allows for flexibility in future changes to the
remapper and avoids various issues with file management.

* Fixes inheritance: now traverses classes at runtime (via cb2obf),
in order to remap methods and fields which were previously missed
due to the lack of this information. Closes #87.

* Faster: highly optimized and with a new compact mapping file.

* Has a cool name

* Uses URLClassLoader, fixing zip file errors. Closes #51.

* Configurable: individual parts of the remapper can be enabled
or disabled, globally or on a per-plugin basis, in bukkit.yml
mcpc.plugin-settings as needed.