Fix remapping Essentials with remap-plugin-file
Re-add the Guava remapping to vcb2obf.srg, so plugins using Guava
can be remapped entirely using remap-plugin-file, and load without
requiring custom-class-loader.

Initially, this mapping was present as of:
 https://github.com/MinecraftPortCentral/MCPC-Plus/commit/e6276e34d56b775247f2f5f8fc8294b6c85e0cf1
but was removed in:
 https://github.com/MinecraftPortCentral/MCPC-Plus/commit/faea2085870fb190cb9f180794c4b159098073bb
since the custom loader at the time always remapped packages:
 https://github.com/MinecraftPortCentral/MCPC-Plus/commit/fd75de507e9ae26ca08a48e2f54ea3cee059065a

Now, as of:
 https://github.com/MinecraftPortCentral/MCPC-Plus/commit/5c0865669c84bc36767362bd89a74e6503d416c9
The custom loader can be completely disabled with custom-class-loader: false,
causing Guava not to be remapped at all, if using remap-plugin-file: true.

This change re-adds the Guava package map to vcb2obf.srg, so Guava-using plugins
can be remapped as before with remap-plugin-file: true.