Fixed chunk unloading issues when using loadChunkOnProvideRequest.
Forge mods require this flag to be turned on during ticket processing in
order to force load chunks. If the flag remains on, many chunks do not get
unloaded properly. To workaround this issue, the ForgeChunkManager will
now turn on the flag for ticket processing then disable it when done.