Fix AbstractMethodError on custom mod recipe classes. Closes #158.
CB changes IRecipe to implement toBukkitRecipe, but mods adding their
own recipe types (for example, IC2 charge-preserving "advanced" recipes)
likely do not implement this method, causing an AbstractMethodError
at runtime when toBukkitRecipe() is called.

If toBukkitRecipe is not defined for the IRecipe, instead of crashing,
now return null as the Bukkit recipe. Note: this could be improved.