Fix crash when placing blocks on RedPower2 timers. Closes #181.
Avoid sending Packet103SetSlot if there is no slot. Right-click on
a timer with a block would appear to place the block on the client,
but the server would recognize the click on a timer and open the GUI
on the client. Interesting logic from vanilla CraftBukkit:

// CraftBukkit - TODO CHECK IF NEEDED -- new if structure might not need 'always'. Kept it in for now, but may be able to remove in future

Added a null check, so now the GUI opens as expected.