Add ComputerCraft fake player to bukkit.yml
This player is used by the ComputerCraft mod when turtles use
turtle.place(). Since ComputerCraft provides a username for this
player, this line is actually unnecessary, but adding it here
is useful for documentation purposes since an administrator
can easily edit this config option to change the name from the
default of "[ComputerCraft]", if needed.