Added new GameRegistry method for registering Bukkit Materials.
In past builds, all modded items/blocks would create a corresponding
bukkit material with the ID passed to the Item/Block class. The material
would be registered without a custom name and use a generic name like "X4077".
This would cause readability issues for plugins that would rely on the
name of a material such as ChestShop.
Now instead of registering a material with a default name of "XID", it
will now use the passed item/block name and use the form "MODID_NAME". If
no name is registered with the item/block, it will default to
"MODID_ITEMID". Keep in mind that the same material name will show up for
blocks that share ID's such as IC2's Solar Panel and Nuclear Reactor. Both
of these will appear as "IC2_BLOCKGENERATOR".