Added new configuration options to enable/disable mod world generators.
Certain mods don't give players the ability to disable their world generators which can cause frustration for players trying to generate a clean world. With this update, as the server starts, all world generator registrations will automatically be added to forge.cfg with the default setting of true. To disable a mod world generator, set the modId to false and restart the server.