Revert to sending block place events in Forge method. Fixes #310, #297.
This reintroduces issue #199, causing mods which override placeBlockAt()
to not send BlockPlaceEvents (same as in builds <=180), including
GregTech machines and RedPower lamps.

However, this fixes: #310 Thaumcraft disappearing items, #297 ICBM launcher dupe,
which were caused by placing duplicate tile entities. The Bukkit
BlockPlaceEvent requires the block is set in the world before the event
is set. If the event is cancelled, the block is reverted. The block
revert was problematic with mod items.

Reverting to the old method for now until a better fix is in place.