Added BlockPlaceEvent for items that bypass onItemUse. Fixes #242
Items such as RP2 microblocks, wires, tubes and BC pipes were bypassing
the onItemUse call causing no block place event to occur. To counter this,
I have added a block place event to World.canPlaceEntityOnSide. As long as
the player passed to this method is not null, the event will be passed.