Pass ClassNotFoundException for compatibility with CommandHelper. See #208.
If the plugin class remapper fails in remappedFindClass(), it will
now throw a ClassNotFoundException, which is passed through findClass().
This allows other classloaders to take a crack at loading the class,
and is the expected behavior for custom classloaders implementing
findClass() - for example, FML does the same, in:

https://github.com/MinecraftForge/FML/blob/master/common/cpw/mods/fml/relauncher/RelaunchClassLoader.java#L185

With this change, the CommandHelper plugin which makes extensive use
of reflection at class load is now able to at least load successfully
(required -XX:MaxPermSize=128M JVM flags however).

Thanks to V10lator of the Surge Dev team for this original fix:
https://github.com/SurgeProject/Surge/commit/7810b5420d0b162777082393b018305231a1c1a2