Added new config option 'dump-materials' to dump material names with id's.
Certain protection plugins such as LWC make use of material names to
decide whether or not to protect a block. In order to use this feature,
the player must know what their material names are.
To enable this new option, set dump-materials to true in bukkit.yml.
All materials will be dumped to the server FML log.
For example, IC2's Solar Generator material name is IC2_BLOCKGENERATOR.