Add additional CraftFakePlayer helper methods
Refactor fake player creation. Fake players can still be created
from mod EntityPlayers but now also with any username:

    getBukkitEntity(World world, String username, boolean doLogin)

getPossiblyRealPlayerBukkitEntity() is similar but will return an
existing Player object if the user is online, otherwise creating a
fake player. This is intended to allow mods to support player-
based permissions for world-altering machines, even when the player
who placed the machine is offline.