Add World trySetBlockAndMetadata for mod event support
Mods can call trySetBlockAndMetadata() instead of
setBlockAndMetadata() etc. to pass a fake player username
and send the appropriate Bukkit events. If the event is
cancelled, the call will return false, and the mod
should not proceed (i.e., not harvest the block).

Lightly tested - subject to change. Tested with block break
events using a quarry in a modified version of Buildcraft.