Add Forge oredict ores to Spigot orebfuscator for #542
Mods with ores registered in the Forge ore dictionary, named
with the "ore" prefix, will be added to the built-in orebfuscator
obfuscate blocks list. This is in addition to the default vanilla
ore blocks.

Note orebfuscator must be enabled for this to have any effect, and
it is off by default (see bukkit.yml).

Current limitation: only ores with block IDs <256 will be obfuscated
because the orebfuscator does not yet fully decode the ID.