Spigot patch 0041: Limit Custom Map Rendering
The default CraftBukkit render sequence for maps is ridiculously slow. By only using it when a custom renderer has been added (rarely in most cases), we can fallback to the Vanilla renderer for general usage. This leads to a much higher effiency overall, especially if no plugins are rendering such maps.

https://github.com/EcoCityCraft/Spigot/blob/d9c76252eb02f7d0d6bea11efe77e3bb1edb0d99/CraftBukkit-Patches/0041-Limit-Custom-Map-Rendering.patch