Fix player for interact events, class cast exception with gravity gun. Fixes #648.
NetServerHandler has two methods for getting the player, for compatibility
reasons. CB had a getPlayer() method returning CraftPlayer, but FML (for
ModLoader compatibility) added a getPlayer() method with the same name but
returning an EntityPlayer.

To resolve this disambiguity (instead of overloading on return type, albeit
possible in the JVM, not in Java source itself), we renamed CB's getPlayer()
to getPlayerB(). But this one usage in PlayerInteractEntityEvent was missed,
and hidden as a compiler warning due to the (Player) cast.

Updating the call to use the correct getPlayerB() method fixes the crash when
triggering player interact events.