Add option to decay flowing lava. Adds MC-4631, #656.
Vanilla has a bug/feature where flowing lava decays
extremely slowly, as in hours/days/never. This change
adds a configuration option to bukkit.yml, which if set
to true (defaults to false for vanilla behavior):

mcpc.flowing-lava-decay: true

will cause a block update to be scheduled after the
random check did not remove the lava, therefore causing
flowing lava to decay at a rate comparable to its normal
flow rate.

For more details see: https://mojang.atlassian.net/browse/MC-4631

This change can be implemented in a mod, however, it requires
either a coremod or jarmod both which are not likely to be
compatible with MCPC+, since CB also edits the BlockFlowing class.
Hence this new config option.