Fix environments on hotloaded dimensions, for Galacticraft. Fixes #709.
Entering the moon dimension in Galacticraft would crash with an NPE
while looking up Bukkit's Environment. Mystcraft dimensions (also hotloaded)
worked fine, because they were loaded in initMystDimension() (for age
support), but other hotloaded dimensions were not. This commit restores
a missed diff in the 1.5.1 update to the vanilla codepath to the
WorldServerMulti() constructor, and allows non-Mystcraft environments.
Now you can fly to the moon.