Disable entity activation range by default
Spigot's entity activation range feature was merged into MCPC+ in:
MinecraftPortCentral/MCPC-Plus@057ee699fed01dd5a41c5d570931d9276605e9d1

This features defines a "range that says at what distance must an entity be
near a player for it to be active."

However, in the world of mods with automation and chunkloaders this
behavior may be unexpected, so this commit disables it by default.

On the other hand, it can "massively improve performance for large servers",
so if you want to turn it on see entity-activation-range* in bukkit.yml.
MCPC+'s default is 0/0/0, vanilla Spigot's is 32/32/16 (animals/monster/misc).

More details on this feature can be found at: http://www.spigotmc.org/threads/entity-activation-range.405/