Fix abstract method error when disabling Netty. Fixes #729.
Disabling Netty (with -Dorg.spigotmc.netty.disabled=true) would fail
with java.lang.AbstractMethodError: net.minecraft.network.TcpConnection.getSocket()Ljava/net/Socket
due to an interaction between FML's runtime deobfuscation and Spigot's
INetworkManager interface changes.

In vanilla (and CB), getSocket() is implemented by TcpConnection,
but Spigot Netty added this method to the INetworkManager interface
which TcpConnection (and NettyNetworkManager) implements. This causes
the getSocket() call to not be remapped since it is invoked through
the interface rather than the class.

To fix this, we remove getSocket() from INetworkManager, and change
NetLoginHandler getSocket() to call the appropriate method through
the appropriate class, explicitly. This preserves the inheritance
hierarchy of the method and allows Netty to be disabled.