Tweak default setting for aggregate-chunkticks
Spigot adds configurable chunk ticking for improved performance.
The default aggregate-chunkticks is well-suited for larger servers,
especially those which choose Spigot over vanilla CraftBukkit for
high-performance. However, on smaller servers this behavior may
be unexpected, leading to confusion when switching from vanilla Forge
to MCPC+.

This commit sets the default aggregate-chunkticks to 0, matching
vanilla behavior. If desired this option can be changed in bukkit.yml,
to Spigot's default of 4 or another value as desired. For more details
on configuring Spigot's features see this guide:
http://www.spigotmc.org/threads/spigot-configuration-guide-last-updated-3-24-2013.802/