Add support for non-enderdragon bosses (e.g. hydra). Fixes #125.
When the Twilight Forest hydra launched its explosive "mortar" attack,
MCPC+ would crash with a ClassCastException while trying to cast the
EntityTFHydra to EntityDragon. This occurred because CraftComplexPart
getParent() hardcodes EntityDragon in getParent().

This commit changes the cast to the more generic Entity class instead
of EntityDragon, adding support for other boss entities. getBukkitEntity()
is implemented in Entity so the same code is called in both cases. The
casting is from the vanilla IEntityMultiPart interface, only implemented by
EntityDragon in vanilla, but in any case will implemented by an
Entity subclass so this cast is safe. Should fix issue #125.