Fix plugin load failure when mods have '+' in their filename. Fixes #748.
When a plugin using ebean (such as SimpleSpawn) loads, it searches
the classpath, including mods. But due to an avaje bug, mod filenames
containing the "+" character (such as Barrels Mod) will be decoded as
a space, and therefore the plugin loading will fail. To fix this we
URL-encode the character before decoding, allowing the plugin to load
successfully. Kinda ugly, open to better fixes..