Fix ObfuscationReflectionHelper remapping obfuscated names. Fixes #704.
We added type info to the field mappings to allow MCPC+ to deobf itself:
https://github.com/MinecraftForge/FML/issues/210 for #643 - but the type is not
available to ObfuscationReflectionHelper; only the field name is passed,
and the type is null.

To allow looking up field names without type, add another mapping entry
with no type. This mapping can possibly be ambiguous, but the API
(getPrivateValue) only accepts the field name, not the type. To lookup
fields with name collisions, the unambiguous srgname can be used --
which always worked with ObfuscationReflectionHelper. As part of this change,
the workaround adding the srgname to FML is removed (but it could be
added back in upstream for version-indep: https://github.com/MinecraftForge/FML/pull/216)

This change fixes mods using ObfuscationReflectionHelper with obfuscated
names, rather than srgnames. It is preferrable for mods to use srgnames
for version-independence, but obfuscated names are valid as well, and are
now properly remapped. Closes #704.