Fix null scoreboard when only overworld is enabled. Fixes #855, #795.
CraftBukkit loops over all of the worlds in MinecraftServer loadAllWorlds(),
and initializes this.server.scoreboardManager for each world (for unknown
reasons, overwriting the previous scoreboard managers).

In MCPC+ we initialize the overworld separately in initOverWorld(), so the
scoreboard manager wasn't initialized if only the overworld was enabled
(e.g., nether and the end disabled and no other multiworld mods/plugins).
This caused NPE's whenever a player did something which affects the scoreboard,
such as taking damage.

To fix this, add the same scoreboard initialization in initOverWorld().