Add support for coremods improperly handling non-existent classes. Fixes #889.
Some coremod transformers will throw exceptions when given a null class,
rather than either returning null or supplying new bytecode to define the
requested class. Since the transformers are executed in a loop, this exception
breaks out of the loop and blocks other transformers from running on the class.

To fix this, we treat an exception thrown from transforming a null class as
if the transformer did not run, therefore allowing others to take a crack at
defining the class. Fixes log spam with CodeChickenCore, closes #889.