Dynamically lookup shaded OBC version for remapping 1.5.2/1.5.1
Adds getNativeVersion() in PluginClassLoader, replacing the hardcoded
'current' field. This method is used to lookup the obfuscation version
(v1_R_3 on 1.5.2, v1_R_2 on 151compat, etc.), and now gets it from the
actual version package in the server jar, rather than hardcoding. This
means it is correct on both 1.5.2 and 151compat builds, where previously
the version was v1_5_R2 in both cases.