Refix issue #120 with inventory pickups not being seen after death. Fixes #965.
The issue is caused by bug in SpecialSource which is explained here :

https://github.com/md-5/SpecialSource/issues/5

Both potionsNeedUpdate and playerInventoryBeingManipulated are being
remapped as 'h' obfuscated. Since EntityPlayerMP uses both of these fields,
setting potionsNeedUpdate causes playerInventoryBeingManipulated to be set
to true. To workaround the issue with SpecialSource, I call 'super'
instead of this so SpecialSource knows we are referring to EntityLiving.