Fixed bug with migrating old MCPC world data to new world structure.
To test if your data moved correctly, check each world_name folder in root
for a DIMX folder. If one exists then the migration didn't work so you
will have to delete the corresponding DIMX folder in world/DIMX. For
example, the old twilight forest data is located in
world_twilightforest/DIM7. The DIM7 folder should be moved to world/DIM7
assuming "world" is the name of your overworld. If you still see a DIMX
folder in world_twilightforest, then it did not move successfully. This
can occur if an existing folder already exists with data in it so simply
remove it. If you need assistance, contact me in #mcportcentral on
irc.esper.net