Added bukkit to forge mappings in bukkit.yml to help with old configs.
If you are using essentials and noticed many player sethomes were no
longer working, it was due to world names being changed to forge. I have
added a bukkit-to-forge mapping section in bukkit.yml to help with this.
Just add the old bukkit world name with the new forge name and it should
work. Towny does not call the same method as essentials so this will not
work for that.