Fix Galacticraft player base server instance patching. Fixes #1050.
Galacticraft, as a coremod, patches attemptLogin() to replace
EntityPlayerMP with its own class. This works since the dev has
explicitly added MCPC+ support but changing the method signature
to fix #756 caused Galacticraft to be unable to find the correct
code to patch.

To solve this, we preserve the original attemptLogin() signature
and method body, adding a new attemptLoginNoEvent() method to
temporarily set an instance variable (vanillaLoginCheck) during
login, rather than passing this information in a parameter. This
allows Galacticraft to find and patch the method, fixing the
"Warning: Could not find player base server instance for player"
spam when right-clicking a rocket.