Special-case Thaumcraft crucible to fix refilling creating a flood. Fixes #471.
Thaumcraft's crucible onBlockActivated always returns false (see MinecraftPortCentral/MCPC-Plus@62890c9b112d887fe919b6da51c99a4c76ae56fa),
rather than returning a flag indicating whether the crucible was filled.
In vanilla Forge, this is no issue, however in CraftBukkit the return value
is used and therefore both the crucible is filled with water, _and_ a water
block is placed on top of the crucible.

To fix this for now, detect whether we're filling a crucible and if so, never
place a water block. Fixes flooding when refilling Thaumcraft crucibles.