Remove logging of broken coremod transformers for now
For diagnosing coremod transformers, #881 added:
"Failed to transform class %s using transformer %s. basicClass = non-null."

But this logging does not occur on vanilla Forge, causing user confusion
when it occurs on MCPC+, even though nothing is different from vanilla and
the failures do not inhibit normal operation of the coremod and server.
Will have to be investigated further in vanilla Forge if this becomes an issue.

Instead, the JVM startup flag -Dfml.debugClassLoadingFiner=true can be used
to log coremod transformations if this information is required.