Fix AIOOB setting item damage on -1 item id for Millenaire. Fixes #1191.
CraftBukkit's ItemStack constructor uses setItemDamage(), but
Forge delegates this to getItem().setItemDamageForStack().. which
requires an Item instance for the item ID. We have a non-itemblock
fallback but since Item.itemsList[-1] is out of bounds, it is not
reached.

To fix this, add a bounds check in getItem(), and return null if
given a -1 item ID. This will trigger the proper non-Item ItemStack fallback.