getSize should check the size of mainInventory directly. Fixes #1252
CraftPlayerInventory.getSize expects to get the size of a player's
inventory excluding armor slots. To do this, they would subtract 4 from
the total size but an issue arises when a mod overrides the
getInventorySize method after extending the equipment slots. This causes
issues with armor since CB assumes the last 4 slots are being used for
armor. To fix this, getSize will now check the length of mainInventory
directly. This should fix issues with Aether and Galacticraft.