Improve error message from invalid MCPC+ installation
If the vanilla Forge jar is (erroneously) extracted into the MCPC+
jar, probably from an ill-fated attempt to update Forge, critical classes
will be overwritten and the server will fail to load. The failure
could occur with NoSuchMethodError(or field/class) anywhere, but our custom
MCPC+ code is first called in the FML relauncher, so if it is overwritten
we'll fail with: java.lang.NoSuchMethodError: cpw.mods.fml.relauncher.FMLInjectionData.obf151()Z

This commit detects this error and shows a message to the user asking
to retest with a clean MCPC+ jar, as a helpful hint in the right direction.