Add stubs for Spigot getIngredients() on Forge oredict recipes
Cannot be implemented properly with the current Spigot API, since
it requires each ingredient must be an ItemStack. However, Forge
oredict recipes allow an "ore name" as the ingredient, a string
corresponding to any number of possible ItemStacks -- for different
"ores" registered by other mods, as crafting-equivalent.

Could possibly be supported in the future if recipe deconstruction
returned a List<List<ItemStack>>, the inner list for each alternative
(thoughts @md_5? @Asphodan?), but for now throwing an UnsupportedOperationException.