Don't skip ticks for entities on edge of chunks. Fixes #1317
AnchorCarts used by Railcraft would halt when reaching a chunk not already
loaded. This was caused by Spigot's updateEntityWithOptionalForce patch
which checks to see if the entity isn't active and if so skips the tick.
Since the tick is skipped, Forge's EntityEvent.EnteringChunk would not be
called and Railcraft would not call ForgeChunkManager.forceChunk.
This fix simply disables the check for entities on a chunk's edge so the
EntityEvent is fired allowing moving entity chunkloaders such as AnchorCart
to work properly.