Revert MinecartContainer/TileEntityChest size to default of 36. Fixes #1262
Apparently, CraftBukkit set the default ItemStack array size of
EntityMinecartContainer and TileEntityChest to 27 instead of 36 since
getSizeInventory() was hardcoded to return 27 in vanilla. While this is
fine for vanilla, it is not for forge mods that may use that extra
space and override getSizeInventory().

Referenced CB commit can be found here
https://github.com/Bukkit/CraftBukkit/commit/27ef303ae40baf15b77ecc7199f701fbfbb3c9ff